refactor: group files by broad categories
