--- src/character-classes/__tests__/encoder.test.ts
@@ -1,12 +1,0 @@
-import { encodeCharacterClass } from '../encoder';
-
-test('buildPattern throws on empty text', () => {
-  expect(() =>
-    encodeCharacterClass({
-      type: 'characterClass',
-      characters: [],
-    })
-  ).toThrowErrorMatchingInlineSnapshot(
-    `"Character class should contain at least one character"`
-  );
-});
