+++ src/components/repeat.ts
@@ -1,6 +1,6 @@
-import type { RegexElement, Repeat, RepeatConfig } from '../types';
-import { EncoderPriority, type EncoderNode } from '../types-internal';
+import { EncoderPrecedence, type EncoderNode } from '../encoder/types';
 import { toAtom } from '../utils';
+import type { RegexElement, Repeat, RepeatConfig } from './types';
 
 export function repeat(
   config: RepeatConfig,
@@ -23,13 +23,13 @@ export function encodeRepeat(
 ): EncoderNode {
   if ('count' in config) {
     return {
-      priority: EncoderPriority.Sequence,
+      precedence: EncoderPrecedence.Sequence,
       pattern: `${toAtom(node)}{${config.count}}`,
     };
   }
 
   return {
-    priority: EncoderPriority.Sequence,
+    precedence: EncoderPrecedence.Sequence,
     pattern: `${toAtom(node)}{${config.min},${config?.max ?? ''}}`,
   };
 }
