+++ src/encoder/__tests__/encoder.test.tsx
@@ -1,6 +1,11 @@
-import { buildPattern, buildRegex } from '..';
-import { one, oneOrMore, optionally, zeroOrMore } from '../quantifiers/base';
-import { repeat } from '../quantifiers/repeat';
+import { buildPattern, buildRegex } from '../..';
+import {
+  one,
+  oneOrMore,
+  optionally,
+  zeroOrMore,
+} from '../../components/quantifiers';
+import { repeat } from '../../components/repeat';
 
 test('basic quantifies', () => {
   expect(buildPattern('a')).toEqual('a');
