+++ src/encoder/types.ts
@@ -1,29 +1,29 @@
-import type { RegexElement } from './types';
+import type { RegexElement } from '../components/types';
 
 /**
  * Encoded regex pattern with information about its type (atom, sequence)
  */
 export interface EncoderNode {
+  precedence: EncoderPrecedence;
   pattern: string;
-  priority: EncoderPriority;
 }
 
 /**
- * Higher is more important.
+ * Order of precedence for regex operators.
  */
-export const EncoderPriority = {
+export const EncoderPrecedence = {
   // Atoms: single characters, character classes (`\d`, `[a-z]`),
   // capturing and non-capturing groups (`()`)
-  Atom: 3,
+  Atom: 1,
 
   // Sequence of atoms, e.g., `abc`
   Sequence: 2,
 
   // Alteration (OR, `|`) expression, e.g., `a|b`
-  Alternation: 1,
+  Alternation: 3,
 } as const;
 
 type ValueOf<T> = T[keyof T];
-type EncoderPriority = ValueOf<typeof EncoderPriority>;
+type EncoderPrecedence = ValueOf<typeof EncoderPrecedence>;
 
 export type EncodeElement = (element: RegexElement) => EncoderNode;
